requirements.txt
fastapi
uvicorn[standard]
gunicorn
pandas
google-cloud-bigquery
python-dotenv
